refactor: :art: extract components from the main component
This changes are in favor of a better code structure
