+++ src/pages/entry/JaavierR/index.astro
@@ -1,10 +1,10 @@
 ---
 import Layout from '@layout'
-import PassGen from '@components/JaavierR/PassGen'
+import PasswordGenerator from '@components/JaavierR/PasswordGenerator'
 ---
 
 <Layout title='JaavierR hacktoberfest 2022 entry'>
-	<div class='h-screen flex items-center justify-center min-h-[25rem]'>
-		<PassGen client:only='react' />
+	<div class='h-screen flex items-center justify-center min-h-[25rem] px-4 sm:px-0'>
+		<PasswordGenerator client:only='react' />
 	</div>
 </Layout>
